Split simulator transport into separate module
Make it easier to reuse the DpeSimulator transport interface without
needing to depend on all the dependencies of the verification tests.
