+++ verification/sim/go.mod
@@ -1,0 +1,7 @@
+module github.com/chipsalliance/caliptra-dpe/verification/sim
+
+go 1.18
+
+replace github.com/chipsalliance/caliptra-dpe/verification/client => ../client
+
+require github.com/chipsalliance/caliptra-dpe/verification/client v0.0.0-20240216212309-5c9f6edaeb55
